strymon_communication: Add source code documentation
